requirements.txt
fastapi
uvicorn
python-multipart     
PyMuPDF              
chromadb            
sentence-transformers
langchain      
google-generativeai      
